Update UI logic to always read from memory
Previously, it was hard to know what the actual state of the game was
via the Crash Helper. If lives were frozen, they could still be changed
via hotkey, without the UI updating. In fact, changing the lives via
hotkey would *never* update the UI. `DataControl` was keeping its own
(often inaccurate) record of the current lives, as was the lives
instance of `GamePointer`, and these two records often did not match.
This set of changes delegates all life and mask count checks to the
`CrashMemory` instance for one unified source of truth.

If caching the values of certain memory lookups is still desired, this
can now more easily be done in `CrashMemory` and synced on the refresh
rate of the UI, as a follow-up change.

Changes:
- `DataControl` now has no knowledge of the current value for lives and
masks and must grab each from `CrashMemory` as they are needed.
- `GamePointer` no longer sets `currentValue` upon a memory write. This
was a large source of inconsistency between UI and game.
- Changes made to tracked memory values by `CrashHelper` now also invoke
the `onValueChanged` callbacks. This means all UI updates can come from
a single source and any future callbacks can respond to all value
changes, both sourced by the game and user. Initially, memory lookup
could be an O(n) operation worst-case, but since it is now always O(1),
the additional memory reads do not add complexity to any simple
functions.
- The `justWritten` property of `GamePointer` has been removed. It was
no longer necessary as a result of the previous change.
- The "Set lives to 0" hotkey no longer overrides the "Freeze lives"
setting. This was a design choice and can easily be reverted, if
desired.
- Lives label on `DataControl` form was moved up by 2 pixels to center
it with the associated buttons.